removed unnecessary driver and changed package name
