--- src/jdaTemplate/app/src/main/java/Bot/SlashCommandListener.java
@@ -1,15 +1,0 @@
-package Bot;
-
-import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
-import net.dv8tion.jda.api.hooks.ListenerAdapter;
-
-public class SlashCommandListener extends ListenerAdapter {
-    @Override
-    public void onSlashCommand(SlashCommandEvent event) {
-        for (Command command : App.commands) {
-            if (command.getName().equals(event.getName())) {
-                command.run(event);
-            }
-        }
-    }
-}
